:sparkles: Actually using the marked items
Evaluation data now includes a mark if an item is consumable or not.
For some reason this causes a small change in the results of the
calculated metrics, but I don't think its significant. I'll keep an
eye on it, but its still possible that this evaluation process
change in the future

Signed-off-by: Antonio Ossa <aaossa@uc.cl>