:zap: Allow models to cache partial results
In recommendaiton mode, at least in CuratorNet, some values can be
cached to improve evaluation performance by avoiding calculation of
repetitive values.

Previously, when a recommendation was generated for a profile across
all possible items, all the items had to be forwarded across the model.
Now a method was added to generate an intermidiate result, to generate
the embedding only once, and that result can be passed as an argument
to the 'recommend' method.

This addition reduced the evaluation script execution time from ~90
minutes to <5. Nice :)

Signed-off-by: Antonio Ossa <aaossa@uc.cl>