:bug: Fix best validation loss initialization
This bug was introduced a while ago and didn't cause any harm, I don't
know why or how. The variable was re-defined every epoch, but in a test
run it caused an error because the variable was not defined. It's
better to add this guard and create the variable explicitly

Signed-off-by: Antonio Ossa <aaossa@uc.cl>