:zap: Calculate metrics using tensors
To improve performance of the evaluation script, some of the methods
in this module were implemented again, using PyTorch tensors. The
main advantage was to avoid or reduce transfers from GPU to CPU while
running the evaluation script, reducing the execution time. It was a
significant speed-up, bringing down the execution time from ~180
minutes to ~95 minutes

Signed-off-by: Antonio Ossa <aaossa@uc.cl>