:zap: Update evaluation script with new features
Previous commits introduced changes to improve the performance (in
terms of the execution time) of the evaluation script. Calculating
metrics using tensors in GPU, reduced the execution time from ~180
minutes to ~95, and using a cached intermidiate result in the
'recommend' method of 'CuratorNet', reduced it to less than 5
minutes. Amazing :tada:

Signed-off-by: Antonio Ossa <aaossa@uc.cl>