:sparkles: Include full filenames in embeddings
First of all, the current state of the repo is working fine, but I
need the full filenames (with extension) to be able to display the
images for inspection.

To visualize items, I need name and extension of the files, which is
problematic only in Pinterest, but could be a problem in the future,
if a new dataset is added.

This commit changes the extract_embedding method too because the items
are marked with the name of each item only, ignoring the extension
(yeah, is not consistent even in Pinterest, but meh). I'll also need
the index to filename maping in the future, so I made the change now

Hope everything goes ok :tada:

Signed-off-by: Antonio Ossa <aaossa@uc.cl>