:bug: Fix calculation of metrics for unseen items
When calculating metrics for datasets with more than one item in the
holdout, it causes the dimensions to mismatch and cause an exception.
This is not relevant when calculating metrics for all the items

I tried different solutions to generalize for dataset with one or
multiple items in the holdout, and now should work. Through expansion
and reshape of the relevant tensors, a shift is calculated (the number
of items) to ignore the already consumed items

It took some time but now it should work :)

Signed-off-by: Antonio Ossa <aaossa@uc.cl>